avoid doing shit for no reason

+++ install.sh
@@ -20,7 +20,12 @@ rm -rf $PYENV_ROOT
 sudo apt-get build-dep qemu
 
 # Python
-curl https://pyenv.run | bash
+export PYENV_ROOT="$HOME/.pyenv"
+export PATH="$PYENV_ROOT/bin:$PATH"
+$PYENV_ROOT/bin/pyenv --version
+if [ $? -ne 0 ]; then
+	curl https://pyenv.run | bash
+fi
 pyenv init -
 pyenv virtualenv-init -
 
@@ -52,36 +57,53 @@ pip3 install pynvim
 
 
 # Neovim, Kubectl
-curl -LO https://storage.googleapis.com/kubernetes-release/release/`curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt`/bin/linux/amd64/kubectl
-curl -LO https://github.com/neovim/neovim/releases/download/stable/nvim.appimage
-chmod u+x nvim.appimage
-chmod u+x kubectl
-sudo mv nvim.appimage /usr/local/bin/nvim
-sudo mv kubectl /usr/local/bin/kubectl
+which nvim
+if [ $? -ne 0 ]; then
+	curl -LO https://github.com/neovim/neovim/releases/download/stable/nvim.appimage
+	chmod u+x nvim.appimage
+	sudo mv nvim.appimage /usr/local/bin/nvim
+fi
+
+which kubectl 
+if [ $? -ne 0 ]; then
+	curl -LO https://storage.googleapis.com/kubernetes-release/release/`curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt`/bin/linux/amd64/kubectl
+	chmod u+x kubectl
+	sudo mv kubectl /usr/local/bin/kubectl
+fi
+
 curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 nvim +'PlugInstall --sync' +qa
 nvim +'UpdateRemotePlugins' +qa
 
 # FZF
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-yes | ~/.fzf/install --all
+which fzf 
+if [ $? -ne 0 ]; then
+	git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+	yes | ~/.fzf/install --all
+fi
 
 #AFL
-wget https://github.com/google/AFL/archive/v2.56b.zip 
-unzip v2.56b.zip
-cd AFL-2.56b && sudo make install && cd ..
-rm -rf AFL-2.56b
-rm v2.56b.zip
+which afl-fuzz
+if [ $? -ne 0 ]; then
+	wget https://github.com/google/AFL/archive/v2.56b.zip 
+	unzip v2.56b.zip
+	cd AFL-2.56b && sudo make install && cd ..
+	rm -rf AFL-2.56b
+	rm v2.56b.zip
+fi
 
 # Shellphish
 pip install git+https://github.com/shellphish/shellphish-afl
 pip install git+https://github.com/shellphish/fuzzer
 
 # Rust
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs > rustup.sh
-chmod +x ./rustup.sh
-./rustup.sh  --default-toolchain nightly --profile complete -y
-rm rustup.sh
+which cargo
+if [ $? -ne 0 ]; then
+	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs > rustup.sh
+	chmod +x ./rustup.sh
+	./rustup.sh  --default-toolchain nightly --profile complete -y
+	rm rustup.sh
+fi
 
 # Switch to fish
 sudo chsh -s /usr/bin/fish $(whoami)
